Initial commit
READnome
By: TVeadedRobots
http://blog.tvheadedrobots.com/

READnome is part of a monomer animation project that I have been working
on. This project makes use of Processing and themonodicc and oscP5
libraries
to turn themonomere into a low resolution display. It will accept any
URL or
file placed in the sketch's data directory and will display it on
thmonomerme.
Each row of tmonomerome represents a single byte on the source file or
URL.
The brightness of the LEDs is also take from the value of the current
byte.
It is an experiment in both low resolution animation and data
visualization.

READnome by TVHeadedRobots is licensed under a Creative Commons
Attribution-ShareAlike 3.0 Unported License.